When adding the person from one of the household records pages, assign them a random password.
This is to prevent validation constraint failure. We'll need to determine the specifics of how this should work at a later meeting.

Presumably, we'll want to check if they're an UMRA member when we're adding them, and then send them a registration confirmation email with a link to set their password.

Signed-off-by: Max Fierke <max@maxfierke.com>